Extract the actions button into separate stateful component
